retool container init and subvolumes
Discovered seveal bugs when converting to use btrfs, as it has lots of problems
doing subvolumes in a chroot.  The new version fixes most of them, but relies on
the fact that the chroot base is its own subvolume.  This is actually good as it
makes cloning tennants and removing them a fairly easy process