example: Hack around NetworkManager / sysfs silliness
in a container sysfs is mounted prior to namespace entry, and so views external
interfaces, rather than those that are in the container. NetworkManager also has
a problem with veth devices as the listed lowerdev is never visible.  So hack
around that some.  We drop network manger and just get the pacakges our network
needs, mount sysfs as part of container boot (so we can see our container
interfaces in sysfs), then modify network-functions to look at the container
sysfs view (and ignore the lowerdev specification)